Military sidebar and other military changes
Also add changes to military behavior.
When the "show military sidebar" option is selected, the sidebar will
display information about the currently selected legion.
Picking a destination for the legion to move to will not cancel its
selection. The only exception is if you move the legion back to its own fort.
If you click a different legion, instead of the currently selected legion
moving there, the clicked legion will be selected.